Type Definition testing_basilisk_runtime::xcm::XcmOriginToCallOrigin

pub type XcmOriginToCallOrigin = (SovereignSignedViaLocation<(ParentIsPreset<AccountId>, SiblingParachainConvertsVia<Sibling, AccountId>, AccountId32Aliases<RelayNetwork, AccountId>), Origin>, RelayChainAsNative<RelayChainOrigin, Origin>, SiblingParachainAsNative<Origin, Origin>, SignedAccountId32AsNative<RelayNetwork, Origin>, XcmPassthrough<Origin>);

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.